refactor: extract shared plan tree-building logic into carina-core
Move duplicated dependency graph and tree-building functions
(build_dependency_graph, build_single_parent_tree, extract_compact_hint,
shorten_attr_name, shorten_service_name) from carina-cli and carina-tui
into a new carina-core::plan_tree module. Both callers now import from
the shared module, removing ~350 lines of duplication.
